Cache refactors
