+++ src/View/Antlers/Language/Runtime/RuntimeParser.php
@@ -28,7 +28,6 @@ use Statamic\View\Antlers\Language\Parser\LanguageKeywords;
 use Statamic\View\Antlers\Language\Parser\LanguageParser;
 use Statamic\View\Antlers\Language\Parser\PathParser;
 use Statamic\View\Antlers\Language\Runtime\Debugging\GlobalDebugManager;
-use Statamic\View\Antlers\Language\Runtime\Sandbox\RuntimeValueCache;
 use Statamic\View\Antlers\Language\Utilities\StringUtilities;
 use Statamic\View\Cascade;
 use Symfony\Component\HttpKernel\Exception\HttpException;
@@ -329,8 +328,6 @@ class RuntimeParser implements Parser
      */
     protected function renderText($text, $data = [])
     {
-        RuntimeValueCache::resetRuntimeCache();
-
         $this->parseStack += 1;
         $text = $this->runPreParserCallbacks($text);
